requirements.txt
redis~=4.6.0
pandas~=2.0.3
pymysql~=1.1.0
requests~=2.31.0
iso8601~=2.0.0
tqdm~=4.65.0
tushare~=1.2.85
python-dateutil~=2.8.1
ta~=0.10.2
akshare~=1.10.43
pytdx~=1.72
tornado~=6.1
cython~=0.29.21
efinance~=0.5.0
multitasking~=0.0.11
rich~=13.3.5
retry~=0.9.2
jsonpath~=0.82
numpy~=1.24.4
sqlalchemy~=2.0.17
timeout~=0.1.2
pydantic~=2.4.2
web3~=6.11.3
func_timeout~=4.3.5
gpt4all~=0.1.7